requirements.txt
requests
beautifulsoup4
watchdog
pyside6
spotipy
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv